python/step6-testing-the-infra-code/pyproject.toml
[project]
name = "my-container-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.172.0",
    "constructs>=10.4.2",
    "pytest>=8.3.4",
]
